+++ Design og implementering/Body SW/Body_main_Version_4/Body_main/Body_main.cydsn/SerialUnit.h
@@ -18,7 +18,7 @@
 
 
 void initUART(void);
-void setdataArray(int ID, int x_data, int y_data, int z_data);
+void setdataArray(uint8 ID, uint8 x_data, uint8 y_data, uint8 z_data);
 void sendDataArray(void);
